StemStudents.java:  I changed the number of students to correspond to the data that we have available.  I also changed the class size to be close to the data, however I wanted even numbers of kids in each class so it is not the exact number from data.  I altered initStudents() to read in from the data file initialInterests.csv which is stored in /data.  I also assign id's to the students based upon the data.
StemStudentsWithUI.java:  I changed the histogram titles to Tech/Eng/Math, Earth/Space Science, Human/Biology to match what we did with the data.  I also changed the # of bins to 7.

TopicVector.java:  Added a new constructor that takes three double values and assigns it to the array.  I also changed the toString to print 4 digits instead of 2.  Used that for some debugging.