I updated the data file initialStudentInput.csv using the new data and analysis from Nancy.
I updated the data file initialActivityInput.csv to change the content of the activities to correspond with the correlation between activities and the three indices (Explore, Science, Human) that Nancy and the team agreed upon.

StemStudents:  Changed the number of students to reflect the new input data file.  Changed the names on the time series to reflect the new index names.  

StemStudentsWithUI:  Changed the names of the charts to reflect new index names.