Feedback loop for activity participation added
Each student now has an array of probability of participating with one for each activity.  I added two functions to Student:  increaseProbOfParticipating() & decreaseProbOfParticipating().  These are just additive functions that use Model.changeParticipationRate parameter (default is 0.05) that has getter/setter and slider created for it.  When willDoToday() is called, it uses the probOfParticipating[] of each student to determine if that student will do a given activity.  Each of the rules need to call the Student.increase/decreaseProbOfParticipating() functions.