First commit since moving to github. Many changes over a long time. Added calibrated model parameters to default.params. Added eighthGradeStudentInput.csv and removed initialInterest.csv (which wasn’t used). Updated ecj.jar and added commons-math and jsc jars. Added an R script for creating the sixth through eighth grade data files. Added KSTest.java which does Kolmogorov-Smirnoff tests. Added some additional param reading to Parameters.java.
Many changes in StemStudents itself. The 8th grade data is now read,
stored, and used for comparison. I put the calibrated parameter values
directly into the code. Added more record keeping. Changed StuffIDo to
participationRates in a few places. Changed the initialization so it can
the numStudents doesn’t have to match the size of the input file
exactly. Added a function for scaling the participation rates, which is
now done after initialization rather than during. This was needed during
the EC calibration. Fine tuned the order of the scheduling levels so
they would run the right amount of time and collect stats at the right
time (step 0, 365, 730, etc.). Made the isSchoolDay/Summer/Weekend
functions public. Added more output. Added code to compare the interest
levels and participation rates of different sets of students using
different metrics.

To StemStudentsWithUI I added several items to the Metrics tab to show
info about participation rates, expected activities per day, etc. In
Student, made the participation rates public, added a function to
calculate the average participation rate. Also made changes to the
parsing to be more robust to weird student IDs, and also to no longer
scale the participation rates during parsing. In RuleSet, made changes
to better handle rule weights. Also made changes to hide the V1 vs V2
stuff, which was probably a mistake in the first place.

Added the tuning directory with three files in it for facilitating the
use of ECJ for tuning this model.